refactor: 캡슐화한 로직으로 의존성 교체 및 리팩토링
